+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Fixed vendor restriction form validation and input handling [#6101](https://github.com/ethyca/fides/pull/6101)
 - Fixed special purpose vendor check in Fides JS logic [#6118](https://github.com/ethyca/fides/pull/6118)
 - Update Add Vendor tooltip to match navigation option [#6111](https://github.com/ethyca/fides/pull/6111)
+- Fixed UX issues with action center tables [#6116](https://github.com/ethyca/fides/pull/6116)
 - Fixed incorrect font styling on some table headers [#6129](https://github.com/ethyca/fides/pull/6129)
 
 ### Removed